Implement Time Unit interfaces for Milliseconds and Nanoseconds

+++ sbk-api/src/main/java/io/sbk/api/RateController.java
@@ -28,6 +28,6 @@ public interface RateController {
      * @param records current cumulative records
      * @param elapsedSec   Elapsed Seconds
      */
-    void control(long records, float elapsedSec);
+    void control(long records, double elapsedSec);
 
 }
